requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref @ file:///croot/asgiref_1724072470684/work
async-timeout==4.0.3
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
channels==4.2.0
charset-normalizer==3.4.1
Django @ file:///croot/django_1731615954020/work
django-sslserver==0.22
djangorestframework==3.15.2
exceptiongroup==1.2.2
google==3.0.0
google-api-core==2.24.1
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.4.0
googleapis-common-protos==1.69.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-core==0.3.41
langchain-text-splitters==0.3.6
langsmith==0.3.11
orjson==3.10.15
packaging==24.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.0a2
pydantic_core==2.29.0
pyparsing==3.2.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse @ file:///croot/sqlparse_1734704446487/work
tenacity==9.0.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
zstandard==0.23.0
